Fix mac build (#691)
Summary:
There were more mac-specific build breakages that needed to be fixed.

One was a relatively simple need to explicitly include `<unordered_map>` due to mac being weird.

The other was a breakage in gtest which does not compile in macos due to an inability in `cmake` to properly set C++17. The fix was to explicitly add `CXXFLAGS="-std=gnu++17"` to force proper compilation, although it may not be a clean fix since the comments indicate that the maintainers wanted to keep `CXXFLAGS` empty.

Pull Request resolved: https://github.com/facebook/redex/pull/691

Reviewed By: yuxuanchen1997

Differential Revision: D35626616

Pulled By: NTillmann

fbshipit-source-id: 5bc3df9f7a5fc7576646279c145895e84d6f895a