don't dedup fillInStackTrace
Summary:
We want to keep the exact source location associated with the top stack-frame of an exception. The strack-trace is captured via direct and indirect calls to java.lang.Throwable.fillInStackTrace. Indirect calls happen through the Throwable constructor. We also need to be careful with outlined calls, as the outlined code might contain such a call to fillInStackTrace.

This diff adds additional safe-guarding that we don't dedup such calls.

Reviewed By: beicy

Differential Revision: D35487946

fbshipit-source-id: 6c548f6fecb085f1455045c10a0a1400085836f0