allow deduping of benign throws
Summary:
Background: We’ve by default disabled deduping of blocks that throw exceptions --- the intention is that the exact source location associated with top stack-frame in the filled-in stack-trace is preserved, and we don't have to guess which one of two possible throwing source locations caused the issue.

We currently avoid deduping *all* blocks that *throw* exceptions; this is overly conservative.

The stack-trace of an exception object is captured by `java.lang.Throwable.fillInStackTrace()`, which is either called indirectly by a constructor of a class deriving
from `java.lang.Throwable`, or possibly via an explicit call.

An explicit throw is in fact benign and can be deduped as long as we don't dedup such a direct or indirect invocation of `fillInStackTrace()` on the exception object flowing into the throw. That's difficult to determine in general, so in this diff we carve out a exception: If the object comes from a non-escaping move-exception, i.e. a re-throw. In particular, that means that fillInStackTrace() cannot have been called (again) after catching and before re-throwing, and the stack-trace remains unchanged. (Except possibly if the exception object is also accessible through some other means, but that is very unlikely (*).)

Reviewed By: yuxuanchen1997

Differential Revision: D35299499

fbshipit-source-id: c0c3738faf0c0684b7490e3a6dabff7cb9476eb5