make large string index into a set
Summary:
It's a bit silly to store the `string_view` (16 bytes) of a DexString as its key in the large string index.
So we don't, and turn the large string index into a set.

If this pattern gets accepted, then I think we can do the same for a few more RedexContext index maps.

A little quirk is that for look-up purposes, we are creating a DexString key that is just "defined enough" to be used as a key into our string set. The provided string does not have to be zero terminated, and we won't compute the utf size, as neither is needed for this purpose.

Reviewed By: ssj933

Differential Revision: D35234572

fbshipit-source-id: a8499f75bc87b9b664f1522c33e021a6d2534d2b