rewrite `compute_locations_closure` cache
Summary:
Instead of caching as we go with heuristics, always compute caches beforehand, and parallelize computation.

This may use a bit more memory, as we conceptually cache everything now (and not just some heuristically determined fraction), but the overhead should be fairly small, as the cache size is just a fraction of already allocated data structures.

This is a behavior-preserving change.

Reviewed By: agampe

Differential Revision: D35804817

fbshipit-source-id: a1407152087b11372a9f5a9655b7481df14e08a2