Aliases for passes
Summary:
Allow a different configuration for a separate runs by introducing alias names for passes.

When the PassManager activates passes, it will check the list of given aliases. If the pass name is an alias, an attempt is made to clone the base pass (this needs support from the base pass, improvements welcome). If that worked, then the new pass is bound to the config of the alias name.

NOTE: Currently each occurrence will create a new alias. Maybe they should be cached instead?

Reviewed By: wsanville

Differential Revision: D35349193

fbshipit-source-id: 223e8894041afa54953545c70ebdfa5f7a823492