allow for block splitting and deduping
Summary:
This code used to look for a pattern where a single block ends in a `throw` instruction. However, via deduping, such a block can be broken up into a chain of blocks connected by simple gotos, still ultimately and unconditionally throwing.

This diff changes the pattern matching to allow for that.

Reviewed By: beicy

Differential Revision: D36418292

fbshipit-source-id: 9b20f1869dbee03ff6928b67bbed8ee520b046fd