discount for unused callee args
Summary:
When an argument to a callee isn't used by the callee, in particular after reducing it based on a call-site summary, then it is likely that some code in the caller that produces the argument will become dead as well.

This change incorporates this idea in the inliner's cost function, applying a cost "discount" for unused callee arguments.

This change does not affect the actual inlining logic, only the cost function.

Reviewed By: beicy

Differential Revision: D36536723

fbshipit-source-id: 81ad677c44c05cfc481ba6067fc126fab9ff9465