Track individual roots in metrics
Summary:
We have similar metrics in the `AnonymousClassMergingPass`. They are missing in the `IntraDexCM` pass.
They help to indicate what kind of classes are we merging.

The `num_roots` metric is similar, but not accurate for interface roots. The reason is that the `roots` in `ModelSpec` can only be classes but not interfaces. The metrics this change adds show a more realistic picture of the number of "roots" being merged as well as what are those "roots" and how big they are (subclass/implementor group size).

Reviewed By: ssj933

Differential Revision: D36673994

fbshipit-source-id: f1d644e5ff27b70cafeab4dfeaaca1ed4e0ca06c