Lowering the default min_counts of IntraDexClassMergingPass
Summary:
`IntraDexClassMergingPass` is the last pass in the Class Merging pipeline. By default we use the `global_min_cout` and `min_count` params to filter out groups of classes that are "too small" to only apply merging to larger groups of anonymous classes.

However, this is not strictly necessary. Now the previous change has been in production for some time, it's safe to lower the default min_count limits. As a result, the pass will merge more anonymous classes and enable larger size savings.

Reviewed By: beicy

Differential Revision: D36611283

fbshipit-source-id: bb7ccac4571444a7c7c13df6c5ab1073f5f26be1