Fix test
Summary:
When running with LocalDce (which is part of the Shrinker, which runs when the InstrumentPass kicks in for *_instrument tests created by D36598576 (https://github.com/facebook/redex/commit/5b3b20a7a24472568330929557505989a93ea84a)), the division-by-zero in the code gets eliminated (thanks to aggressive LocalDCE implementation since 2015: D2385863).

This makes the test so that the division result is used, and the division-by-zero is triggered.

Reviewed By: ssj933

Differential Revision: D36795364

fbshipit-source-id: bff43c4873d4270cfc47ab339c4b107210db50c5