use editable cfg in DexAssessments.cpp
Summary:
Use editable cfg in DexAssessments.cpp. When editable cfg is built, duplicated positions are introduced. Therefore, in this diff, we don't measure duplicated position like:
assume b1->b2. if b1 has the last position last_pos, we just need to check if any position before the first insn in b2 is the same as last_pos. if so, that position in b2 will be marked as duplicated and won't be counted.

However, the above algorithm doesn't cover all cases, so we still expect larger metrics number than before.

Reviewed By: NTillmann

Differential Revision: D36041157

fbshipit-source-id: 7e8cf21a7914ac794e9d17fd30d42a8cfa262d79