Disable for classes with external field refs
Summary:
`UnwrappableHttpEntityWrapper` inherits from Apache's `HttpEntityWrapper`, which is external to the APK, and contains a protected field, `wrappedEntity`.

Since `ClassSplitting` currently moves methods to classes without the same inheritance chain, splitting out methods referencing `wrappedEntity` cause an exception at runtime in Messenger.

It seems `get_visibility_changes` should prevent this, but it currently does not consider references to fields in external classes to count as having the potential to cause visibility changes. (It also doesn't consider unresolvable fields to potentially cause visibility changes either - not sure if that's a concern in reality, but seems wiser to avoid allowing class splitting in these cases.)

Added a new function `has_unresolvable_or_external_field_reference` to block class splitting if an unresolvable or external field reference is found in a class' method.

Reviewed By: agampe

Differential Revision: D36709469

fbshipit-source-id: afe9c87f041fd82c80c55c2ebb0e7dce0e27f796