Option to propagate sufficient hotness/warmness to cold callees
Summary:
When `enable_hotness_propagation` is enabled in the outliner (or partial application pass) config, this function will propagate sufficient "hotness" or "warmness" to callees of sufficiently hot and warm methods whose entry blocks were executed according to block profiles.

This is to mitigate the fact that method profiles track appearances and call counts for physical methods which exist at the end of a dyna build of an app, and we have no precise means of attributing these appearances/call counts to methods which exist earlier in the IR, and which are later inlined and only "executed" via their inlined blocks.

For example, if, in a dyna build, foo is inlined into bar, we will only have method profile info for foo in a resulting profile if it's executed, and bar will appear to not be executed. In a regular/optimized build using this profile, the outliner may then outline from bar, which will later be inlined into foo, causing an outlined method call to appear in a sufficiently hot method.

In practice, this causes many outlined functions to show up in method and block profiles for cold start, which isn’t intended.

It works by first identifying all methods, m, which were not called according to method profiles, but whose entry blocks were executed according to block profiles (i.e. with first source blocks in their entry blocks which were executed).

It then takes this list of candidates, and marks them warm or hot if they had a caller which was warm or hot.  Since possibly warm or hot methods could be called by other possibly hot, the code also iterates until a fixpoint is hit.

Disabled everywhere for now.  May enable in future if it can be proven to make a difference in an A/B.

Reviewed By: NTillmann

Differential Revision: D36459637

fbshipit-source-id: 8e9378c531064670b2bedee6e16d45eda3c7e20d