Wrapper program to zip and dump .arsc file
Summary:
A silly program to take an .arsc file, put it into a zip file (making sure it is named resources.arsc) and call `aapt d --values resources` on it.

We seem to perform zipping/unzipping of .apk files from python later in Redex, not C++ (do we have any zip creator libraries in C++?) so this will get called in the later diff from a C++ test.

Reviewed By: thezhangwei

Differential Revision: D36451952

fbshipit-source-id: 01d54818665826e85bb906697f5db60de542a0f4