TableSnapshot convention to unfork more ResourceTypes.cpp code, make edit operations cleaner
Summary:
Today, `android::ResTable` internal details are consulted to iterate through resource ids, resolve values, etc. This is not a great API for us to be building upon, because it is stateful (the unforked version of this would require setting the current "configuration" to access all resources for a given ID).

The following diffs in the stack will move away from `android::ResTable` to our separate APIs. Anything that uses `android::ResTable` will move over to `TableSnapshot` as its replacement.

`TableSnapshot` is really meant to not do a whole lot, it mainly has a TableEntryParser and all of that data, plus it inflates the actual string pool classes for the encountered headers (plus some convenience methods).

`TableSnapshot` should ideally not leak out into optimization passes; that is a sign of a pass that is not well abstracted for .apk inputs and .aab inputs. It may however be useful for test cases that really want to verify nitty gritty things.

Reviewed By: ssj933

Differential Revision: D36337347

fbshipit-source-id: e11ac80f137b0d12af93e1eb018b9915180aba1c