Reimplement get_resource_strings_by_name
Summary:
The pre-existing `get_resource_strings_by_name` implementation is pretty sketchy, due to the stateful nature of of `android:ResTable`. The API really doesn't let the caller express the intent, since it relies on whatever the state of `android:ResTable` is, to return the default config or most representative value for the case in which a resource has multiple values in different configs (like in the supplied test case).

IMO we should not even have a function like this in our codebase, but for historical reasons the only reasonable-ish implementation I could think of would be to make it return values in all configs (including chasing references like the old impl did) since this thing already returned a vector.

Reviewed By: ssj933

Differential Revision: D36337388

fbshipit-source-id: 66ef55c62e7b33899629a2ae462fc12dc3d5af78