Use aapt d in ResourceSerializationTest
Summary:
Remove the last use of `android::ResTable` so we no longer have to import that code into our project. This diff converts the test to do its validation via `aapt d --values resources` and ensure the return code is 0.

`aapt d` under the hood this tool is just gonna create/parse a `android::ResTable` and run the print function on it, so effectively it should achieve the same thing we want without us needing to actually import the source code.

Reviewed By: ssj933

Differential Revision: D36451958

fbshipit-source-id: 7020daa31b737f9eaa2ccf55c875b224afddfedf