Creating Java Profile Parser to help with Instrumentation Test
Summary:
We have written testcases and asserted that particular basicblocks were hit. However, we had hardcoded these blocks by reading the metadata file and source block file too. To avoid the need of reading the files, we are writing a Java Parser that will automatically read the necessary profile files and can be queried whether a particular block is hit.
- Wrote a java file MetadataInfo.java that saves the information parsed from redex-instrument-metadata.txt
- Wrote a java file MetadataParser.java that parses the necessary files and provides functions to Test java files to use
- Created a startup function that executes before every jUnit test and loads everything from the Metadata file and source dictionary

Reviewed By: ssj933

Differential Revision: D37120594

fbshipit-source-id: 3380360c9c7edce45c5ed75ed87ab602c57623ff