Redo testcases by using the MetadataParser
Summary: In D37120594, we created a Java Profile Parser that can take the metadata and parse the profile array for the correct bitvector based on the function name. We updated all the Junit tests by removing the hardcoded values and replacing it with functions calls to MetadataParser to provide the hitValue based on function name, stats array, and block name.

Reviewed By: ssj933

Differential Revision: D37191885

fbshipit-source-id: 9efacc2a13ea31b7449bb7da34e33a89f70d9995