enable MSO in staging/stable
Summary:
Enable MSO in staging/stable. The size win is between 431.3 and 444KB compressed (see BSBs). See https://fb.workplace.com/groups/RedexTeam/permalink/1128864164327686/ and https://fb.workplace.com/groups/RedexTeam/permalink/1128126201068149/ for more information.

Also enable it for Workplace for Android, as it uses the same method profiles.

Will write a post to announce that we enable the optimisation in stable and will link it here, in the description.

Reviewed By: NTillmann

Differential Revision: D37278823

fbshipit-source-id: 2ee84b8bf9a9dbd13b0cad1e093334e38fd8ac22