Reset elements of ArrayNullnessDomain when passed through calls
Summary:
There's a subtle bug in GlobalTypeAnalysis when tracking the nullness for the elements of an array. When an array is passed through a call, if the array is modified in the callee, the nullness domains of the array elements may not be properly updated in the caller environment. This change fixes the bug including the followings:

- Always reset array elements NullnessDomain when the array is passed to a call as an argument
- New integration test cases
- Add more tracing

Reviewed By: ssj933

Differential Revision: D37337596

fbshipit-source-id: d802f9031e7d1198b0d2fe52c1ef1660b3396ab0