Deduplicate encode/decode for map and set.
Summary:
This diff stack intends to deduplicate map and set to the extent possible; currently they are nearly copy-pasted which makes improvements more tedious and error-prone. We already have one place where they have diverged behaviorally.

None of the diffs have change in external behavior, only internal rearrangements. The net LOC is about -400. Interestingly, the mere act of simplifying and merging these various functions gives a ~1.1% latency improvement - mostly due to the 6th diff.

It may be helpful to peek at later diffs in the stack to see why a particular change was made, for fuller context.

This first diff just merges and simplifies the common codec functionality. It is now the central place where `IntegerType` is defined and asserted to be valid.

Reviewed By: yuxuanchen1997

Differential Revision: D37281004

fbshipit-source-id: a961add7a9621d61563ed3eaba19a826121138bf