preshrink
Summary:
Instead of initially shrinking several copies of a root method, we only have to (pre)shrink it once.
This makes the pass run faster.

The code here gets a little bit more complicated, but the next diff in the stack simplifies it again. This intermediate diff is mainly so that...
This is a behavior-preserving transformation.

Reviewed By: thezhangwei

Differential Revision: D37403437

fbshipit-source-id: acfd48ac00c1973161ef2d0cc2daeba8b4388bea