consider several reduced variants
Summary:
Before, when the initial static analysis determined that a root method had the potential to inline allocations of N different types, then it was an all-or-nothing situation: Either all of them could be successfully inlined, or none. As inlining can fail for various obscure reasons (e.g. cross-dex references), this was unfortunate, as not inlining any particular type means that it can't be fully eliminated.

This is getting improved here, but not just giving up, but trying to find a *subset* of inlinable types. We do not try all possible subsets to avoid an explosion in the number of reduction attempts, but instead have a heuristic that will try N subsets if we have N inlinable types to begin with.

This does not affect the actual logic of inlining and thus the correctness of the pass, but it's rather a twist to the heuristic / cost function.

Reviewed By: thezhangwei

Differential Revision: D37338595

fbshipit-source-id: 9b38691be6ac2fc6a114d1759e27c1699c5c5c01