rewrite of SignedConstantDomain, use less memory
Summary:
This greatly simplifies the internal representation and computation of joins and meets:
- We keep a boolean to remember if we excluded zero (nez)
- We keep two `int64_t` values to represent the lower and upper bound of the tracked interval (possibly still to be understood as excluding zero, if zero would be within the range, but the nez flag is set)

Plus vtable pointer makes 32 bytes, half of what it used to be.

This is a behavior-preserving change; the public interface of `SignedConstantDomain` does not change.

With this change, all members of Redex' `ConstantValue` `DisjointUnionAbstractDomain` are now at most 32 bytes, bringing down the size of `ConstantValue` from 96 bytes to 48 bytes.

Reviewed By: agampe

Differential Revision: D33012933

fbshipit-source-id: 2b18b5cbf72f982953a6cc71bead8a8a8ba7a625