Reduce mem usage by wrapping sets in HashedSetAbstractDomain
Summary:
This is a behavior-preserving change.

We wrap the `unordered_set` into a `unique_ptr`, so that `sizeof(HashedSetAbstractDomain)` shrinks from 80 to 32 bytes, and only non-empty domain values take a hit.

This is important when the domain is used in a `DisjointUnionAbstractDomain` whose size is dominated by the largest member (see in particular Redex' `ConstantValue`).

Reviewed By: yuxuanchen1997

Differential Revision: D37157233

fbshipit-source-id: c41e4066a7788adc47e5357944bc0d6b9b84e819