use less memory
Summary:
This is a behavior-preserving change.

We wrap the value into a `unique_ptr`, so that `sizeof(ObjectWithImmutAttrDomain)` shrinks from 64 to 24 bytes, and only non-empty domain values take a hit.

This is important when the domain is used in a `DisjointUnionAbstractDomain` whose size is dominated by the largest member (see in particular Redex' `ConstantValue`).

Reviewed By: thezhangwei

Differential Revision: D37157235

fbshipit-source-id: d61fe8109a9acd03f34615973df1e02a697da47f