Merge various update functions.
Summary:
This is the most complicated diff in the refactor. It deals with the fundamental operation of updating a leaf node's value.

Map internally relies on `update`, which ultimately delegates to `combine_leaf`/`combine_new_leaf`. Set has `insert` and `insert_leaf`, which are identical except the latter starts with a leaf already allocated.

All of these are merged into a more general `update_leaf`. This is one of two overloads, either taking an existing leaf to update or taking the key of where a new leaf would go. Both also take a leaf node operation callback.

The leaf node operation callback is invoked with the current leaf (if any), and returns a proposed new value for the leaf. The return type can be any of: a value, an optional value, or a leaf node intrusive pointer. This is statically checked - this is both convenient to use and helps the inliner eliminate irrelevant branches without having to peek through a variant discriminant.

From `update_leaf`, the auxiliary functions follow easily.

 ---

Map's use of `update` actually required `std::function` overhead as well as allocation of a temporary auxiliary leaf node to do an insert. In contrast, the new `update_leaf` abstraction requires no forced allocation and encourages reuse of leafs slightly more often (e.g., now in `use_available_value` we can directly reuse a leaf instead of copy a value). Although behaviorally identical, this diff has latency improvements.

Reviewed By: NTillmann

Differential Revision: D37281009

fbshipit-source-id: a3c293cb5b9b911c695549ec4a537a0cd19669d3