Remove ADL ambiguity in `PatriciaTreeCore`.
Summary:
This happens because the free functions implementing PatriciaTreeCore accept `boost::intrusive_ptr` as arguments, so functions in the boost namespace are candidates at the callsite. It didn't occur pre-refactor because these functions took a concrete `std::function` argument that caused disambiguation. We removed the overhead of `std::function` but the result is purely templated functions.

To remove the ambiguity this ensures each of these functions contains the word "tree", "leaf", or "key" somewhere in the name, since these (currently) don't exist anywhere in boost. While we're here, also use the "_by_key" suffix consistently for functions that locate things by key.

The long-term solution would be to wrap the intrusive_ptr in some manner to disable ADL, but this is much easier.

Reviewed By: thezhangwei

Differential Revision: D37542177

fbshipit-source-id: 05775537b7146fb9a531305127d06ca82b23ae2f