Fix nondeterminism in NormalizeConstructor
Summary:
# Context
When dedupping ctors for anonymous classes, we could run into situations where updating the ctor signature causes method identity collision. We rely on `type_reference::fix_colliding_dmethods` to address the issue. However, the list of colliding ctors we pass into the function is unordered. That seems to be the source of nondeterminism detected in the IntraDexClassMergingPass late, after we expanding the merging targets.

# Mitigation

- Replace unordered set/map with ordered counterparts in `dedup_constructors` for `DexMethod*`
- Make sure that the `colliding_methods` list passed to `type_reference::fix_colliding_dmethods` is ordered

Reviewed By: agampe

Differential Revision: D37543736

fbshipit-source-id: eb383db099edbcd764e8885bc927dd0b11a71979