custom graph implementation for AliasedRegisters
Summary:
This uses less memory and is faster.

The main insight to achieve this is that alias groups are relatively rare, while the number of vertices increases linearly with the code size, which includes the number of registers used. We need to query alias groups, but we don't actually need to know which (singleton) vertices exist. Thus, we move out the machinery to track vertices into a shared data-structure, and only keep the relatively sparse edge information for each node.

To achieve this, we introduce the `AliasGraph`, a custom bidirectional graph with edges over `Value`. The graph provides exactly those operations needed by `AliasedRegisters`. In particular, each vertex can have at most one out-edge. The out-edge points to the representative of an alias group, and in-edges point to all other members of an alias group. The graph and all copies of it will share one underlying `VertexValues` mapping, a fast bi-directional mapping between `Value` and `vertex_t` indices.

The existing code in `AliasedRegisters` is adjusted to reflect the migration from the general `boost` `adjacency_list` to the new more limited `AliasGraph` API, and so to avoid ever querying all vertices, which isn't actually needed, as all operations only actually care about non-singletons, i.e. alias groups represented by edges.

The new approach is faster, because...
- It avoids copying the vertex mapping in the vast majority of cases
- It changes formerly linear look-up operations find to constant time.

This is a behavior-preserving change.

Reviewed By: beicy

Differential Revision: D37585051

fbshipit-source-id: 325749025ac4a1ed6558244408acf9d0daea20ad