Fix bug in order
Summary:
The BSB failures on D37675533 show that `less_than` is not an order. (It's obviously not a strict order, but it's also not a non-strict order, but I don't think it was ever meant to be a non-strict order, as that just wouldn't make sense.)

Fixing this is important to avoid non-deterministic behavior at best, and build breaks at worst when `std::sort` would notice. I noticed the issue while working on a performance improvement that brought out unexpected non-determinism (D37662008 V3).

This is getting fixed here.

The new implementation of handle_insert_order_at_merge is also faster, going from O(N^2 log N) to O(N^2), by moving repeated work from the less_than to doing it only once upfront.

Reviewed By: beicy

Differential Revision: D37678679

fbshipit-source-id: a0882b33f448c7af776c317acf5931b6ac17e3de