Require integer-like elements in BitVectorLattice.
Summary:
The new optimizations require specializing on the lattice element being an integer-like type, one-to-one with the naturals.

We could do fancy template work to transparently opt in to these optimizations, and still have a fallback general case for the other non-scalar types...but in practice everything is already a zero-based enum. And I suspect we would strongly suggest this approach anyway, avoiding the fallback cases. So this diff tightens the requirements to require a strictly continuous zero-based integer-like element set. If we feel the need to loosen this later it is easy to do so.

This removes the need for allowing specification of Hash/Equal. However, removing them fully is a backward-incompatible change so for now they are merely ignored.

Differential Revision: https://www.internalfb.com/diff/D37775181?entry_point=27

fbshipit-source-id: 8b23355cbda650ab449d97fb7faa4b7ad7fb965f