Minor shuffle of BitVectorLattice.
Summary:
The upcoming diffs add a lot more implementation inside the `fad_impl` namespace. This ends up putting the actual useful public class, `BitVectorLattice`, way down at the bottom and hard to find.

This rearranges so that the public interface goes first.

 ---

The goal of this stack is to speed up BitVectorLattice encode & decode operations by ~6x-16x. The first three diffs are some refactoring and restriction so the implementation details are less entangled with their use. Then the performance improvements are added.

Differential Revision: D37775182

fbshipit-source-id: 49395fa292f3ac34ed7ce985a3dbfe721c3856a7