Use clz for decoding.
Summary:
By relabelling the lattice elements during encoding, we can ensure each encoding is uniquely identifiable by the number of leading zeros it has.

This gives a cheap way to map an encoding to an index (i.e., to the relabelled element), which we can then un-relabel back to the original element for decode by lookup table.

There are a few cascading benefits of this transformation. For example, top and bottom always have a fixed final encoding and can be made constexpr rather than dynamically found. In turn, most functions can be made static constexpr, allowing deeper inlining of operations where one argument is top or bottom.

Differential Revision: https://www.internalfb.com/diff/D37775179?entry_point=27

fbshipit-source-id: be8382c3300db79da09a382ffd0b5162b32e3b9b