Simplify BitVectorLattice.
Summary:
Currently, when BitVectorLattice uses the opposite lattice sometimes it implicitly converts the encoding back to the lower lattice and sometimes it does not. This is a leaky abstraction where opposite lattice has to understand how it's used to work.

This diff makes BitVectorSemiLattice completely uniform, and BitVectorLattice only uses semi-lattice opaquely. This is required before BitVectorSemiLattice can dictate its own encoding and operation.

Differential Revision: D37775180

fbshipit-source-id: a10392ff968d96e50abaf5c761a5df3dcbad1c00