Make exception throwing cold.
Summary:
`BOOST_THROW_EXCEPTION` (and by extension `RUNTIME_CHECK`) expands to a significant amount of code, responsible for allocating, copying, and destroying exception vtables and strings. This prevents even tiny functions from inlining since the callee stack space is desirable for optimizing this process.

We expect to never throw. This diff hoists the act of constructing and throwing an exception behind a cold noinline function - this minimizes pollution of the calling site and allows these functions to freely inline.

Reviewed By: yuxuanchen1997

Differential Revision: D37623344

fbshipit-source-id: 33b4fd7b24ad94469a3e3bfd5d8fdd1edd50985b