Check if SwitchMap lookup table is actually populated
Summary:
There is a rare case when removing the SwitchMap lookup table for switch/when statements. The look up table may be created but not populated with actual cases by the <clinit> of the generated class and, therefore, never referenced anymore.

This could be the outcome of a preceding optimization. We should check instead of fail the map lookup.

Reviewed By: beicy

Differential Revision: D37834631

fbshipit-source-id: deece5d8848d156dcd2b799144e488a170985a8b