Move define_type to ResourceTableFile abstraction
Summary:
Both of these sub classes of ResourceTableFile work in generally the same way, they collect info about edits that are pending, which take effect during serialization.

Move an existing collection to the base class, and continue in that fashion for the definition of types. A subsequent diff will make this actually take effect for .aab files.

Reviewed By: ssj933

Differential Revision: D37507466

fbshipit-source-id: 6be81687ffecb40404a556489f802e6b061489ae