Instrument Non-Loop Blocks at the End of Method
Summary:
After D37623677 scheme of instrumenting every block had a huge overhead, we implemented a hybrid version. We added onBlockHit in every loop block due to our analysis that loop blocks are at max ~15% of the codebase. On the other hand, we create a bitvector of hits for non-loop blocks and just loop over it at the end of the method. This scheme gives precise hit count information compared to D37477836 and has less overhead than D37623677.
- Updated BlockInstrument to instrument the onBlockHit function into only loop blocks and create a bitvector and pass it to onNonLoopBlockHit so it can increment the count for non-loop blocks

Reviewed By: agampe

Differential Revision: D37672623

fbshipit-source-id: 07f1696ab3c55929ce629c05bde8975e8ec94e5e