Implementing Loop-Based Block Hit Counting
Summary:
Redex's BlockInstrument pass already checks if a block is hit or not. Now, we are updating it to add the functionality that will count the number of times a block is hit. To offset the overhead, we only need to add new instrumentation for blocks within loops and we can stick to the normal hit array used before.

- Updated BlockInstrument to check whether a block is within a loop or not to instrument the onBlockHit function
- Updated MetadataInfo and MetadataParser to parse the metadata file correctly since the addition of the three more columns needed for block hit counting

Reviewed By: agampe

Differential Revision: D37477836

fbshipit-source-id: ad3711592a10ea945fcddbfce878b0f8894680f1