New BlockInstrument changes to take into account Synchronized Methods
Summary: In D37500185, we noticed the newly added Hit Counting had issues in synchronized methods as seen in P512516091 that came from https://fburl.com/code/oyl08hxd. The issue was an imbalance in monitor stack so we changed the hit count code by splitting the block and creating a new block just for OnBlockHit.

Reviewed By: agampe

Differential Revision: D37477076

fbshipit-source-id: bb7d71e0052146f3cb146b528c5ee63f55105595