Instrument every Block with OnBlockHit
Summary:
After D37477836 and D37477076 scheme for counting Basic Blocks did not have a huge overhead, we implemented a version that put onBlockHit in every block to check the overheads for this implementation. This scheme gives precise hit count information compared to D37477836. Due to now instrumenting every block, we had to make the hitStats array bigger so we changed from int (4 bytes) to short (2 bytes) because we do not need fine granularity for our ideas for optimizations.

- Updated BlockInstrument to instrument the onBlockHit function into every block
- Updated MetadataParser to fix the hitOffset due to the code now instrumenting every block and to change type of the hitStats array from int to short

Reviewed By: agampe

Differential Revision: D37623677

fbshipit-source-id: c0ee1768eed681fa42de861d935bfea2fda44460