support for "incomplete"ly inlinable types
Summary:
Before, the ObjectEscapeAnalysis would only attempt to eliminate allocations of a type if *all* new-instance instructions of that type could be fully tracked, and have been found to not escape (making them an "inline anchor"). This is getting relaxed here, by now also considering individual new-instance instructions, as long as the new object doesn't escape in some local context. Only a local cost analysis is made for such cases. Considering all such cases fully would take a lot of time, as we'd try to fully inline and shrink each involved root method. To keep the overhead down, we make a rough estimation of the expected size delta after inlining, and discard early those cases which are highly unlikely to result in a size improvement.

If a method has inline anchors of N > 1 types, we still only consider N subsets, which could be improved on.

Reviewed By: agampe

Differential Revision: D37887087

fbshipit-source-id: f20f31d2436353f5c13d44f660116a28f5a18156