Start to implement a Rust library
Summary:
This diff is to start the crate repo and maybe also gauge interest among current SPARTA users. We will need a highly performant abstract interpretation library in Rust for our work on new analyses. Here I am only creating notion of an `AbstractDomain` and some of the basic data structures. But at a high level, we need several things to have a working abstract interpreter:

- Data structure: Patricia Tree. This might be a little intimidating. But figuring out the way it works and re-implementing it in idiomatic Rust will be well worth it IMO. Technically we can use `HashMap` for the interpretation semantics to be correct. But Patricia Tree has proven to be the major performance win in the C++ implementation.
- WeakPartialOrdering: Technically fixpoints will work on any ordering. WeakPartialOrdering is the latest iteration to enable efficient (in fact the most efficient way possible) Parallel fixpoint iteration. Will need a graph trait for objects we would like to iterate on.
- MonotonicFixpoint: Core part of the algorithm.
- Parallel version of the MonotonicFixpoint: This might be a little tricky because we would like to leverage Rust's data race safety while still maintaining the high performance of the current C++ implementation.

I admit the code in this particular diff was initially written long before this diff existed as my practice of learning Rust back in 2020. It could be wrong in many ways. Comments are welcome.

FAQ:
- Why not FFI to C++?
I think we could consider using FFI to bridge to C++ for complex but self-contained components like the patricia tree. The real answer is that I am still evaluating the complexity of FFI as opposed to redoing the implementation in (potentially Unsafe) Rust, while my philosophy remains as keeping the implementation of the library simple and straightforward in idiomatic Rust.

Differential Revision: D38364639

fbshipit-source-id: 4c14e53636ad52435f4cd2f2c1358e4b5d931ceb