Basic Patricia Tree
Summary:
Patricia Tree implemented in Rust. I did not attempt to copy the algorithm from the C++ implementation.

This diff adds a basic Patricia Tree by its definition in Rust. Copying of such structures is essentially free thanks to its immutable internal structure. This is not a performance optimized version of it and it may fall short of what C++ can offer. However, its in *almost completely* safe Rust, in which implementing the Core logic of insertion and iteration takes < 200 lines of code.

Currently the key is constrained to 64 bit integral values. However, as long as 1) we fully implement the heap allocated BitVec and 2) the used defined types are convertible to and from BitVec, I don't see why this isn't applicable to all types of normalized, bit-representable data.

There is a slight bit of unsafety here when accessing union fields in BitVec. We don't absolutely need it but it's reasonable to consider that we can potentially do heap allocated BitVec to represent arbitrarily deep trees.

A few TODO items that we can slowly add to the code. in the order of importance:
- Being able to tell whether `insert` operation has actually performed insertion or performed update.
- Removing node from tree.
- Implement `SetOps` traits to use with Abstract Domain. Specialize it with merge tree/intersect tree algorithm.
- Investigate threading use: apart from internal counters of Rc, nothing should be mutable.
- IntoIterator for Set and Map interface types

Differential Revision: D38449054

fbshipit-source-id: 30e81e23519defa1036b64cd5748097e07c475de