WPO implementation part I.
Summary: This is the part I of WPO implementation, which adds most type definition and functions. The main missing part is the building function of the order. Note that in the C++ version of SPARTA, it's implemented in WpoBuilder, but here we prefer implementing that in a single function (with multiple closures for utility functions). It'd be a long function but makes more sense logically.

Reviewed By: yuxuanchen1997

Differential Revision: D38613114

fbshipit-source-id: 5a57ef834158b70e5d86ad8a7810c461197e9170