Propagate exceptions from tasks to the main thread
Summary:
Currently, if an exception is thrown within a task run in `SpartaWorkQueue`, it will crash its thread, and everything will continue as normal. The main thread that called `run_all` will not know about the exception.

This diff adds propagation of any exception thrown in a task to the main thread. This is usually the intended behavior since this is the sign of an error in the program. If not desired, one can simply add a try/catch in the task body.

Reviewed By: dkgi

Differential Revision: D38788124

fbshipit-source-id: 067d1c5c5be8cd4e168f5201c84f592a4918b9e4