WPO implementation part II.
Summary:
This diff adds the implementation for Weak Partial Ordering. There are some differences between the paper and the C++ implementation. I tried to make this as close to the C++ version as possible to make sure its correctness (but in Rust style). Later we can make more improvements to this leveraging Rust features, if necessary.

A new function `size` is added to the Graph trait to get the number of all nodes. This is used by the union find structure when performing DFS. I expect that this function could be easily implemented by any actual graph type.

There isn't a very good union find Rust crate. Currently, for quick prototype, I use the one from petgraph, which is an actively maintained crate. I will think more about if we should implement it directly in our own library.

Regarding testing WPO, this diff does not include all the tests available in the C++ repo. I found there are plenty of duplicate code in that test file. So I decided to add them later in a separate diff, with some common functions/macros to reduce the duplicate code.

Reviewed By: yuxuanchen1997

Differential Revision: D38847698

fbshipit-source-id: df7feaf7ba3faaee72f63d97dad03e7a9a566df8