WELL_KNOWN_METHODS
Summary:
This follows the same pattern as WELL_KNOWN_TYPES. We can migrate other uses of `DexMethod::get_method` later, and possibly get rid of some one-off caching in static variables.

The immediate benefit is for DedupBlocks, where I recently added a `DexMethod::get_method` call in a somewhat hot path.

This is a behavior-preserving change.

Reviewed By: beicy

Differential Revision: D38988124

fbshipit-source-id: c547e047b7927db6f02500d7610d5795931565f2