fix inlining when a method contains an invoke-super and an invoke-virtual to the same true-virtual method ref
Summary:
A while ago, D29698629 (https://github.com/facebook/redex/commit/83746061f00e77a3f76483f71e651d922b7cd6a4) sneaked in a bug, by adding the target of an invoke-super to the true-virtual method space. This could then cause other invoke-virtuals (when occurring in the same method as the invoke-super) with the same method-ref to be wrongly considered for inlining.

And even if not inlined, it could cause subsequent control-flow to be interrupted by a `throw null` if the target method has no return, even though other overridden methods have a return.

All this is getting fixed here. The core change is to have `MultiMethodInliner::get_callee` not return a callee if it's exclusively associated with true virtuals.

Reviewed By: thezhangwei

Differential Revision: D38958426

fbshipit-source-id: a2a8ce2ef0f6e4fc10f6588b16e02c9381501e8a