HashMapAbstractPartition
Summary:
Implemented `HashMapAbstractPartition` in Rust. Similar to C++ version's `HashedAbstractDomain` with similar semantics.

I used generic associated type in the `AbstractPartition` trait. It probably doesn't need to be there though. I originally wanted some kind of higher kinded type argument for the enum. But it turns out to be unsupported and probably won't be worthwhile doing.

Reviewed By: NicholasGorski

Differential Revision: D39121960

fbshipit-source-id: 775ff22a2d837c65e8a65ff55d2a62eb6e710eaa