adding "huge~methods" metric
Summary:
This corresponds to huge methods that the SplitHugeSwitchPass is trying to break up.

This part of the assessor can run with or without editable cfgs being build; in the case of editable cfgs, they no longer contain `OPCODE_GOTO`s, so we take a special step to make sure we account for those.

Reviewed By: agampe

Differential Revision: D39123499

fbshipit-source-id: 70f7a7b6eba96ee5806217a4946df5b1e839b305