Removing `max_split_methods` option
Summary: Instead of "preallocating" a certain number of method refs, we just scan the dexes and figure out how many method refs are still available. In practice, this is a large number, much larger than the small previous preallocated number of method refs. Thus the benefits are better InterDex class-to-dex allocation, and more method refs available to the SplitHugeSwitchPass. (Note that the InstructionSequenceOutlinerPass also doesn't preallocate any method-refs, but instead scans for available method refs.)

Reviewed By: agampe

Differential Revision: D39120259

fbshipit-source-id: 2c44309cd6c4c638aab320b701c2ded9e0b4f2f0