fix application on direct methods
Summary:
The logic of the pass would allow for direct methods to be split up, but then it would generate the wrong opcode, which blew up on diffs further up the stack. This is being fixed here.
Also, we now never create virtual methods, which could lead to wrong overrides in new virtual scopes.

Reviewed By: agampe

Differential Revision: D39121257

fbshipit-source-id: 8de727ab90ba506919ebe5da7d6e06a73e60b33e