Avoid unnecessary usage of `std::function`
Summary:
`std::function` is slow. There are many blog posts out there to establish a case against it. e.g. [this one](https://blog.demofox.org/2015/02/25/avoiding-the-performance-hazzards-of-stdfunction/)

I don't think many of our `std::function` usages are justified. There might be a case for the stored lambda in the WPO, but otherwise, we shouldn't use it.

The downside of this diff is that it may cause the error messages to be a bit more obscure when passing a predicate that is of the wrong function type. Good news is that it seems possible to improve with C++20 Concepts.

Reviewed By: NicholasGorski

Differential Revision: D39366517

fbshipit-source-id: 859d7ac8f58512ffdcd0250b5083dadbb4976038