Remove unnecessary destructors.
Summary:
Destructors remove implicitly-defined move constructors, but we often use them for `static_assert`. The lack of move constructors shows up as a latency cost in certain analysis.

Using a destructor for this purpose is typically unnecessary, so this diff removes destructors where possible. The only case it's actually needed is to defer the static assert until all the involved types are complete. For this we define a private empty base class to contain the assertions, which still allows the deferred but guaranteed static assert behavior. (The empty base class is therefore the unmovable one, but it's empty so we don't care.)

For the sake of smaller diff size this leaves virtual base classes with no members as is, as well as classes where the absence of a move constructor doesn't matter in practice.

Differential Revision: D39355689

fbshipit-source-id: 8dfb0d24f71566878fc032965fd28bbf9c664256