Include name mapping in cross dex minimizer json.
Summary:
The JSON primarily works over abstracted/simplified identifiers. This allows one to map back to the original underlying type/field/method name.

The resulting JSON was already somewhat large before, and now it is even larger. However, given that this is all opt-in there isn't a particularly strong reason to make the content optionally granular.

Reviewed By: NTillmann

Differential Revision: D39317551

fbshipit-source-id: f91cc9add94e34dd768aea82af617508427d26ce