lose some informational content
Summary:
As the original code would only ever record a field once, `num_field_references` ends up being the recorded field count.
The `per_cls_refs` data was only ever used in level-3 tracing, and ever there was dubious, as only the first reference of a field would get recorded. So I am removing it here.

This is in preparation of the next diff up the stack, which makes things concurrent.

This is a behavior-preserving change.

Reviewed By: wsanville

Differential Revision: D39408407

fbshipit-source-id: cc60b837614d55214fb0231cf11b6acb1338b142