Insert check-cast for replaced interface return type
Summary:
When the return type of a method is declared as interface, the type check on the returned type is more relaxed. It is treated as an assignment to interface. However, if we replace the interface return type with a class, we might introduce a type error on the return.

This change fixes the bug in `SingleImplPass`:
- Collect return instructions assigning to an affected return interface type
- Insert `check-cast` for the collected return
- Add a test replicating this bug
- Revert blocklist edit in D39218405

Reviewed By: NTillmann

Differential Revision: D39480291

fbshipit-source-id: 498b1ffff296f30eb727e241f3fae28af61045f8