impl Fn -> impl FnOnce in AbstractEnvironment::update
Summary:
I don't think we need `impl Fn` for `op` parameter in `AbstractEnvironment::update`. Logically, in `update` function, we try to get the state of `variable`, if one is found, we call `op` to update its value, it's called only once. And currently all implementations of `update` use FnOnce in fact.

Change to FnOnce will allow us to move captured variables at the caller site, which is important.

Reviewed By: yuxuanchen1997

Differential Revision: D39578669

fbshipit-source-id: 8f6e0e35d3b93e45575fadd8419882fce8cba560