Explicit cast to make Json library happy again (#722)
Summary:
It seems like in jsoncpp library, there are some ifdefs around the constructor of `Json::Value` that take a 64-bit integer (stemming from https://fburl.com/code/wf6652ze or something like that).

IDK why that has to happen, but in this case type/field/method refs are never gonna be that big so just cast to int should be fine right?

Pull Request resolved: https://github.com/facebook/redex/pull/722

Reviewed By: agampe

Differential Revision: D39581005

fbshipit-source-id: ebd6fbfe459a6e9b893301729a1225fddc14c8df