have get_entry_state_at / get_exit_state_at returns a const-ref
Summary:
This eliminates a copy, and should make things faster.
For some domains, the copy is inconsequential. But for others, in particular largish `HashedAbstractPartition` values, it can be significant.

For IPCP (and GTA), the (unnecessary) copying effectively made `compute_entry_state` be O(n^2) where n is the number of out-edges (tracked invoke instructions) in any method. This diff makes it linear, as one would hope for.

While IPCP (and GTA) benefit the most from this change, all algorithms using the `MonotonicFixpointIterator` should benefit a bit.

Reviewed By: arnaudvenet

Differential Revision: D39513978

fbshipit-source-id: 161210c685e4164c120524c218d26ee087c734bb