Fix SwitchEquiv
Summary:
Add a helper for the first insertion position of source-blocks (they should not lead to a split of move-result-ish operations).

Use this, and the newer `Block`-based `insert-before` in `SwitchEquivFinder` to correctly insert copies into leaves. This avoids the wrong insertion *after* an existing leading source-block.

NOTE: SwitchEquivFinder is still troubling me. If the branches are not otherwise removed, or if the whole transformation is stopped, then we will have the copies still around, and that will matter for counting.

Reviewed By: NTillmann

Differential Revision: D39833183

fbshipit-source-id: 2ae51c2a6adface6c5a0f741d5ccf7e027d1e8a3