Cache intraprocedural analysis results
Summary:
Here, we track all WholeProgramState "dependencies" in analyze_node, and store the result and dependencies in a cache. In future runs, we check if we find a matching input and dependencies, and then reuse the result. New metrics show that we get cache hits in the vast majority of cases.
The cache increases memory usage, but not beyond the recently achieved memory savings in IPCP.

This is a behavior-preserving change.

Reviewed By: beicy

Differential Revision: D39489704

fbshipit-source-id: d3f99aa106b7ede714ae20f44906ee2a352a2f73