Default refine_to_external to true
Summary:
This is motivated by the discussion in https://github.com/facebook/redex/issues/724.

Resolving pure refs including external ones is necessary for optimizations involving reference update, like AnonymousClassMerging. I don't see a risk of doing this, since we do have fall back logic in `ExternalRefsManglingPass`. When the `min_sdk` level is not supported or if the `android_sdk_api_file` is missing, we fall back and disable the option.

Reviewed By: NTillmann

Differential Revision: D40310534

fbshipit-source-id: f2356014711a832026b6b70ed51e35d84e15e6eb