Environment captures state after the current instruction
Summary:
The environment of the analysis captures the register to abstract domain mapping at each program location (at each instruction). However the current logic maps an instruction to the state before the execution of that instruction. This can be confusing sometimes, since when we collect all the reflection sites in a method, we will miss the instruction performing the actual reflective operation in favor of the following one.

For instance, a `INVOKE_VIRTUAL v0 Ljava/lang/Object;.getClass:()Ljava/lang/Class;` is a reflective instruction or a reflection site. The current logic will miss this instruction in the collected reflection sites, andl include the following `MOVE_RESULT_OBJECT` instead as the 1st reflection site. This is counterintuitive. This change fixes the issue.

Reviewed By: NTillmann

Differential Revision: D40449395

fbshipit-source-id: 97917d03dbb12255a4359594f8e7108a6d79d4c5