option to concurrently explore a set of cross-dex-ref-minimization alternatives
Summary:
This introduces a new optional strategy for cross-dex-ref-minimization.

Within each dex, it basically as before: Our strategy for picking the next class to emit in a dex:
- at the beginning of a new dex, consider a "seed" class, i.e. a class with a high seed weight
- otherwise, pick the "best" class according to the priority scheme that prefers classes that share many applied refs and bring in few unapplied refs; do until dex is full

What's new here is that for each dex, we have the option to explore a set of alternatives in parallel, choosing the "best" one to continue with. We use a cost function that tries to minimize the remaining "difficulty" of assigning the remaining classes to dexes, derived from the remaining references, where more isolated references (used in fewer classes) have a higher weight (I went through a lot of different variations here; this one is just the best overall cost function I could find).

This is a behavior-preserving change, as the new option is off (set to 1) by default.

Reviewed By: NicholasGorski

Differential Revision: D39874197

fbshipit-source-id: 907c22626f443fdc31ed13fd022bef292bfb382c