fix secondary profile comparator logic when one method is not in stats_map
Summary: I found a bug in the case when methods were being compared and one isn't found in the profile.  In this case we defaulted to initial ordering, but errors in the sort order would cascade.  For example, if `a` has appear_percent 0.0 and `b` isn't found, we can end up saying that `a < b`, which can result in live code mixed in with dead code.

Reviewed By: NTillmann

Differential Revision: D40600877

fbshipit-source-id: 29eece4d873d5a2ee0eaccc16e8622943dd6a691